+++ CHANGELOG.md
@@ -59,7 +59,10 @@ The format is based on [Keep a Changelog].
   `describe-face` (which delegate to `completing-read-multiple`
   internally) now use Selectrum by default. To select additional
   candidates within a supported command, use `TAB` and input
-  `crm-separator` (`,` by default). See [#53], [#80], [#74].
+  `crm-separator` (`,` by default). See [#53], [#80], [#74].  The user
+  option `selectrum-completing-read-multiple-show-help` can be used to
+  control display of additional usage information in the prompt
+  ([#130], [#132]).
 * We provide a `selectrum-completion-in-region` function now and
   install it on `completion-in-region-function` in `selectrum-mode`,
   so `completion-at-point` will use Selectrum when there is more than
@@ -137,6 +140,23 @@ The format is based on [Keep a Changelog].
   keyword argument `:may-modify-candidates` to re-enable the old
   behavior for cases where it is safe and the performance gains are
   useful. See [#74].
+* Working with the default candidate has been improved in cases where
+  it is not in the candidate list. Such candidates are currently shown
+  in the prompt message. For example, the command `lgrep` might
+  suggest searching through files matching `*.el` instead of just a
+  specific file. See [#120], [#122].
+  * While there is no user input, the default candidate remains
+    visible in the prompt message. Previously, it would be hidden when
+    the prompt line was not selected. Unchanged is the behavior is to
+    hide the default candidate when text is typed, so that it is only
+    visible when it can be submitted (similar to the effect of
+    `minibuffer-electric-default-mode`).
+  * The default candidate shown in the prompt message is now displayed
+    with the face `selectrum-current-candidate` when it is selected.
+  * Now that Selectrum always shows the default candidate when it can
+    be submitted, it now attempts to remove the default candidate from
+    prompt messages that already contain it. This decreases
+    redundancy.
 
 ### Bugs fixed
 * You can now use the undo system in the minibuffer. Previously,
@@ -173,6 +193,9 @@ The format is based on [Keep a Changelog].
   ([#107], [#108]).
 * You can now select and submit empty input and for file prompts
   existing paths when require-match is non-nil ([#67], [#125]).
+* The default candidate is now first selected, even when it is not in
+  the candidate list, conforming with expectations. Previously, the
+  first candidate in the list was selected instead. See [#120].
 
 [#4]: https://github.com/raxod502/selectrum/issues/4
 [#12]: https://github.com/raxod502/selectrum/issues/12
@@ -219,9 +242,13 @@ The format is based on [Keep a Changelog].
 [#108]: https://github.com/raxod502/selectrum/pull/108
 [#113]: https://github.com/raxod502/selectrum/issues/113
 [#118]: https://github.com/raxod502/selectrum/pull/118
+[#120]: https://github.com/raxod502/selectrum/issues/120
+[#122]: https://github.com/raxod502/selectrum/pull/122
 [#125]: https://github.com/raxod502/selectrum/pull/125
 [#126]: https://github.com/raxod502/selectrum/issues/126
 [#127]: https://github.com/raxod502/selectrum/pull/127
+[#130]: https://github.com/raxod502/selectrum/issues/130
+[#132]: https://github.com/raxod502/selectrum/pull/132
 [raxod502/ctrlf#41]: https://github.com/raxod502/ctrlf/issues/41
 
 ## 1.0 (released 2020-03-23)